NEW: 简单生成技能特效的GameOBject

+++ Unity/Client/Assets/GameScripts/SKillSystem/Config/LogicFrameConfig.cs
@@ -13,7 +13,5 @@
     /// <summary>
     /// 逻辑帧间隔 （毫秒）
     /// </summary>
-    public static float LogicFrameIntervalMS = 66f;
-    
-    
+    public static int LogicFrameIntervalMS = 66;
 }
